mavlink: imported new version with MEMINFO message
git-svn-id: https://arducopter.googlecode.com/svn/trunk@3205 f9c3cf11-9bcb-44bc-f272-b75c42450872

+++ libraries/GCS_MAVLink/include/protocol.h
@@ -45,16 +45,16 @@
 MAVLINK_HELPER mavlink_status_t* mavlink_get_channel_status(uint8_t chan);
 #if MAVLINK_CRC_EXTRA
 MAVLINK_HELPER uint16_t mavlink_finalize_message_chan(mavlink_message_t* msg, uint8_t system_id, uint8_t component_id, 
-						      uint8_t chan, uint16_t length, uint8_t crc_extra);
+						      uint8_t chan, uint8_t length, uint8_t crc_extra);
 MAVLINK_HELPER uint16_t mavlink_finalize_message(mavlink_message_t* msg, uint8_t system_id, uint8_t component_id, 
-						 uint16_t length, uint8_t crc_extra);
+						 uint8_t length, uint8_t crc_extra);
 MAVLINK_HELPER void _mav_finalize_message_chan_send(mavlink_channel_t chan, uint8_t msgid, const char *packet, 
 						    uint8_t length, uint8_t crc_extra);
 #else
 MAVLINK_HELPER uint16_t mavlink_finalize_message_chan(mavlink_message_t* msg, uint8_t system_id, uint8_t component_id, 
-						      uint8_t chan, uint16_t length);
+						      uint8_t chan, uint8_t length);
 MAVLINK_HELPER uint16_t mavlink_finalize_message(mavlink_message_t* msg, uint8_t system_id, uint8_t component_id, 
-						 uint16_t length);
+						 uint8_t length);
 MAVLINK_HELPER void _mav_finalize_message_chan_send(mavlink_channel_t chan, uint8_t msgid, const char *packet, uint8_t length);
 #endif // MAVLINK_CRC_EXTRA
 MAVLINK_HELPER uint16_t mavlink_msg_to_send_buffer(uint8_t *buffer, const mavlink_message_t *msg);
